% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_Beta.R
\name{double_Beta}
\alias{double_Beta}
\title{Double Beta}
\usage{
double_Beta(n, y, p, alpha1, alpha2, beta1, beta2, a)
}
\arguments{
\item{n}{number of successes}

\item{y}{number of attempts}

\item{p}{proportion of each beta distribution in mixture}

\item{alpha1}{alpha hyperparameter for first beta distribution}

\item{alpha2}{alpha hyperparameter for second beta distribution}

\item{beta1}{beta hyperparameter for first beta distribution}

\item{beta2}{beta hyperparameter for first beta distribution}

\item{a}{the alpha value used to compute the tail probability (interval)}
}
\value{
a plot of the prior density, a plot of the posterior density, and a list of the point and interval estimates
}
\description{
Double Beta
}
\examples{
double_Beta(10, 4, 0.5, 2, 8, 8, 2, 0.025)
}
