% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stan_model.R
\name{run_stan_model}
\alias{run_stan_model}
\title{Run Stan Model}
\usage{
run_stan_model(n, y, p, alpha1, alpha2, beta1, beta2, chains = 3, iter = 10000)
}
\arguments{
\item{n}{number of successes}

\item{y}{number of attempts}

\item{p}{proportion of each beta distribution in mixture}

\item{alpha1}{alpha hyperparameter for first beta distribution}

\item{alpha2}{alpha hyperparameter for second beta distribution}

\item{beta1}{beta hyperparameter for first beta distribution}

\item{beta2}{beta hyperparameter for first beta distribution}

\item{chains}{number of models to run}

\item{iter}{number of iterations for each model}
}
\value{
Histogram of the posterior and tabular summary of the MCMC posterior sample
}
\description{
Run Stan Model
}
\examples{
run_stan_model(10, 4, 0.5, 2, 8, 8, 2, 3, 10000)
}
