% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_Beta.R
\name{single_Beta}
\alias{single_Beta}
\title{Single Beta Function Plot}
\usage{
single_Beta(n, y, alpha, beta, a)
}
\arguments{
\item{n}{the number of trials run for the given experiment}

\item{y}{the number of successes}

\item{alpha}{the alpha hyperparameter for the beta distribution}

\item{beta}{the beta hyperparameter for the beta distribution}

\item{a}{the alpha parameter for determining the tail probability (interval)}
}
\value{
a plot of the prior, likelihood, and posterior densities
}
\description{
Single Beta Function Plot
}
\examples{
single_Beta(n=10, y=4, alpha=2, beta=8, a=0.025)
}
